backend/requirements.txt
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
Pillow==8.3.2
djangorestframework-simplejwt==5.0.0
pyjwt==2.3.0
django-cors-headers==3.10.0
psycopg2-binary